+++ Source/BuildSync.Client/Source/Forms/PeersForm.cs
@@ -19,18 +19,12 @@
   3. This notice may not be removed or altered from any source distribution.
 */
 
+using BuildSync.Core;
+using BuildSync.Core.Utils;
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
-using BuildSync.Core;
-using BuildSync.Core.Utils;
 using WeifenLuo.WinFormsUI.Docking;
 
 namespace BuildSync.Client.Forms
@@ -74,7 +68,7 @@ namespace BuildSync.Client.Forms
 
             MainListView.ListViewItemSorter = ColumnSorter;
         }
-        
+
         /// <summary>
         /// 
         /// </summary>
@@ -152,8 +146,8 @@ namespace BuildSync.Client.Forms
                 Item.SubItems[2].Text = StringUtils.FormatAsTransferRate((long)Peer.PeakRateIn);
                 Item.SubItems[3].Text = StringUtils.FormatAsTransferRate((long)Peer.AverageRateOut);
                 Item.SubItems[4].Text = StringUtils.FormatAsTransferRate((long)Peer.PeakRateOut);
-                Item.SubItems[5].Text = StringUtils.FormatAsSize((long)Peer.TotalIn);
-                Item.SubItems[6].Text = StringUtils.FormatAsSize((long)Peer.TotalOut);
+                Item.SubItems[5].Text = StringUtils.FormatAsSize(Peer.TotalIn);
+                Item.SubItems[6].Text = StringUtils.FormatAsSize(Peer.TotalOut);
                 Item.SubItems[7].Text = Peer.LastSeen.ToString("dd/MM/yyyy HH:mm");
                 Item.SubItems[8].Text = StringUtils.FormatAsSize((long)Peer.TargetInFlightData);
                 Item.SubItems[9].Text = StringUtils.FormatAsSize((long)Peer.CurrentInFlightData);
